backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.124.4",
    "motor>=3.7.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.6",
    "uvicorn>=0.38.0",
]
